% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bm_prep}
\alias{bm_prep}
\title{Read in and wrangle data from Blue Maestro sensors}
\usage{
bm_prep(filename)
}
\arguments{
\item{filename}{path to the csv file obtained from the blue maestro app.}
}
\description{
Read in and wrangle data from Blue Maestro sensors
}
