% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bm_summary}
\alias{bm_summary}
\title{summarise blue maestro data}
\usage{
bm_summary(l, exclude = NA, start_date = NA, end_date = NA, min_t = NA)
}
\arguments{
\item{l}{a list of csv filenames point to raw blue maestro data}

\item{exclude}{which sensor numbers to exclude}

\item{start_date}{exclude measurements before this}

\item{end_date}{exclude measurements after this}

\item{min_t}{exclude rows with temperatures below this (typically for error exclusion)}
}
\description{
summarise blue maestro data
}
