% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_layers}
\alias{create_layers}
\title{Create Topographic Wetness Index and assocated layers}
\usage{
create_layers(dem, fill.sinks = TRUE, deg = 0.1)
}
\arguments{
\item{dem}{A digital elevation model loaded in using the Raster package}
}
\description{
Thanks to https://stackoverflow.com/questions/58553966/calculating-twi-in-r
}
