% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_es}
\alias{get_es}
\title{Saturation pressure}
\usage{
get_es(temp_c)
}
\arguments{
\item{temp_c}{temperature}
}
\description{
Thanks https://physics.stackexchange.com/questions/4343/how-can-i-calculate-vapor-pressure-deficit-from-temperature-and-relative-humidit
}
