% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_folded_aspect}
\alias{get_folded_aspect}
\title{Calculate folded aspect}
\usage{
get_folded_aspect(aspect)
}
\arguments{
\item{aspect}{Aspect in degrees. Can be a raster or a vector of numbers}
}
\value{
folded aspect, ranges from 0-180, with 180 being the hottest aspect (SW)
}
\description{
Calculate folded aspect
}
