% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_hli}
\alias{get_hli}
\title{calculates Heat Load Index (McCune and Keon 2002)
Code adapted from K.C. Rodman}
\usage{
get_hli(dem)
}
\arguments{
\item{dem}{a digital elevation model in SpatRaster format}
}
\description{
calculates Heat Load Index (McCune and Keon 2002)
Code adapted from K.C. Rodman
}
