% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_latitude_raster}
\alias{get_latitude_raster}
\title{get a raster of latitude values from a dem without reprojecting}
\usage{
get_latitude_raster(dem)
}
\arguments{
\item{dem}{a digital elevation model in SpatRaster format}
}
\description{
get a raster of latitude values from a dem without reprojecting
}
