% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_twi}
\alias{get_twi}
\title{Calculate Topographic Wetness Index and upslope area}
\usage{
get_twi(dem, deg = 0.12, fill.sinks = T, resolution = NA)
}
\arguments{
\item{dem}{A digital elevation model loaded in using the Raster or terra package. Units must be in meters}

\item{deg}{some kind of degree thing}

\item{fill.sinks}{if true, small depressions are smoothed out in the dem}

\item{resolution}{NA is the default, assuming the native XY resolution of the
raster is in meters. Set to the resolution in meters if the native xy resolution
of the raster is in degrees}
}
\description{
Thanks to https://stackoverflow.com/questions/58553966/calculating-twi-in-r
}
