% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_vpd}
\alias{get_vpd}
\title{Vapor pressure deficit}
\usage{
get_vpd(rh, temp_c)
}
\arguments{
\item{rh}{relative humidity}

\item{temp_c}{temperature}
}
\description{
Thanks https://physics.stackexchange.com/questions/4343/how-can-i-calculate-vapor-pressure-deficit-from-temperature-and-relative-humidit
}
