% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hli}
\alias{hli}
\title{Heat Load Index}
\usage{
hli(dem, check = TRUE, force.hemisphere = c("none", "southern", "northern"))
}
\arguments{
\item{dem}{terra SpatRaster class object}

\item{check}{(TRUE/FALSE) check for projection integrity and calculate central latitude for non-geographic projections}

\item{force.hemisphere}{If country is split at the equator, force southern or northern hemisphere equation c("southern", "northern")}
}
\description{
calculates Heat Load Index (McCune and Keon 2002)
This used to be a part of this package, but we discovered that it's avalable in
spatialEco (Evans 2023), so we just have it as a wrapper now, so as not to
ruin any scripts.
}
\references{
Evans JS, Murphy MA (2023). \emph{spatialEco}. R package version 2.0-2,
\url{https://github.com/jeffreyevans/spatialEco}.
}
