% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mef_station_data}
\alias{mef_station_data}
\title{Weather Station Data for Manitou Experimental Forest}
\format{
\subsection{'mef_station_data'}{

A simple feature collection with 59 features and 8 fields
\describe{
\item{temperature_c}{2 meter air temperature in degrees Celsius}
\item{humidity_pct}{Relative Humidity at 2 meters}
\item{timestamp_ts}{datetime, at one hour increments}
\item{vpd_kPa}{Vapor pressure deficit in kilopascals at 2 meters}
\item{ymd}{date: year month day}
}
}
}
\usage{
mef_station_data
}
\description{
Weather Station Data for Manitou Experimental Forest
}
\keyword{datasets}
