% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotwise_summary}
\alias{plotwise_summary}
\title{summarise blue maestro data to the plot scale}
\usage{
plotwise_summary(bm_df)
}
\arguments{
\item{bm_df}{a data frame created by the bm_prep function}
}
\description{
summarise blue maestro data to the plot scale
}
