% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotwise_summary_std}
\alias{plotwise_summary_std}
\title{summarise blue maestro data to the plot scale, standardized against a single time series}
\usage{
plotwise_summary_std(mdf, stdf)
}
\arguments{
\item{mdf}{a data frame created by the bm_prep function}

\item{stdf}{data frame consisting of a single time series of vpd, rh and temperature against which to standardize

data('mef_micro_data'); data('mef_station_data')
plotwise_summary_std(mef_micro_data, mef_station_data)}
}
\description{
columns of standardized values start with 's' (e.g tmax standardized is stmax)
}
