% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_sp}
\alias{predict_sp}
\title{apply predictions of a spatial process model to a raster}
\usage{
predict_sp(rast_stack, spmod)
}
\arguments{
\item{rast_stack}{a raster stack with all the predictor variables}

\item{spmod}{a spatial process model object}
}
\description{
apply predictions of a spatial process model to a raster
}
