% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_envfit}
\alias{tidy_envfit}
\title{convert vegan::envfit output to a tidy data frame}
\usage{
tidy_envfit(x)
}
\arguments{
\item{x}{the output of the vegan::envfit function}
}
\description{
convert vegan::envfit output to a tidy data frame
}
