% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_sp}
\alias{tidy_sp}
\title{Tidily create a spatial process model}
\usage{
tidy_sp(sf_df, vars, y, ...)
}
\arguments{
\item{sf_df}{sf-style point layer with predictor and response variable}

\item{vars}{character vector of predictor variables}

\item{y}{name of response variable}
}
\description{
Tidily create a spatial process model
}
