% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twi}
\alias{twi}
\title{Calculate Topographic Wetness Index and upslope area}
\usage{
twi(dem, deg = 0.12, fill.sinks = T, resolution = NA, include_upslope_area = F)
}
\arguments{
\item{dem}{A digital elevation model loaded in using the Raster or terra package. Units must be in meters}

\item{deg}{some kind of degree thing}

\item{fill.sinks}{if TRUE, small depressions are smoothed out in the dem}

\item{resolution}{NA is the default, assuming the native XY resolution of the
raster is in meters. Set to the resolution in meters if the native xy resolution
of the raster is in degrees}

\item{include_upslope_area}{if TRUE, outputs upslope area and TWI, if FALSE (default), output is only TWI}
}
\description{
Thanks to https://stackoverflow.com/questions/58553966/calculating-twi-in-r
This function requires topmodel, which is no longer on cran. It is forked at admahood/topmodel/topmodel
}
\examples{
require(elevatr)
require(sf)
require(terra)
pt <- data.frame(lat = 39.16456241323459, lon = -105.34602042323105) |>
 st_as_sf(coords = c('lon', 'lat'), crs = 4326)

rst <- elevatr::get_elev_raster(locations = pt, z = 12, prj = 4326)
# utm zone 13
# rst1 <- terra::project(as(rst, "SpatRaster"), method = 'bilinear', y = st_crs(26913)$proj4string)

# either specify resolution and use original lat long raster,
# or use a dem in meters
twi <- topomicro::twi(rst, resolution = 14.78322,
                     include_upslope_area = FALSE)
 plot(twi)


}
