% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotwise_summary}
\alias{plotwise_summary}
\title{summarise blue maestro data to the plot scale}
\usage{
plotwise_summary(mdf)
}
\arguments{
\item{mdf}{a data frame created by the bm_prep function}
}
\description{
summarise blue maestro data to the plot scale
}
\examples{
data('mef_micro_data')
plotwise_summary(mef_micro_data)



}
