% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotwise_summary_std}
\alias{plotwise_summary_std}
\title{summarise blue maestro data to the plot scale, standardized against a single time series}
\usage{
plotwise_summary_std(bm_df, std_df)
}
\arguments{
\item{bm_df}{a data frame created by the bm_prep function}

\item{std_df}{data frame consisting of a single time series of vpd, rh and temperature against which to standardize}
}
\description{
columns of standardized values start with 's' (e.g tmax standardized is stmax)
}
